requirements.txt
torch
transformers>=4.35.3
accelerate
pillow
numpy
scipy
pandas
statsmodels
tqdm
